--- client/pages/travelDestination/index.module.css
@@ -1,98 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  /* justify-content: center; */
-  width: -webkit-fill-available;
-  height: -webkit-fill-available;
-  min-height: 100vh;
-  padding: 0 0.5rem;
-  overflow: auto;
-}
-
-.title {
-  width: 600px;
-  height: 200px;
-  font-size: 70px;
-  text-align: center;
-}
-
-.image {
-  width: 300px;
-  height: auto;
-}
-
-.inp {
-  padding-left: 3%;
-  border: none;
-  border-left: 1px solid #b6b5b5;
-}
-
-.inp:focus {
-  outline: #e95464;
-}
-
-.inp[type='text']::placeholder {
-  color: #999; /* プレースホルダーの色を薄いグレーに設定 */
-}
-
-.subject {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100px;
-  height: 100%;
-  text-align: center;
-}
-
-.searchBox {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 500px;
-  height: 300px;
-  border: 1.5px solid #e95464;
-  border-radius: 30px;
-  box-shadow: 0 15px 15px -10px rgb(0 0 0 / 20%);
-}
-
-.box {
-  position: relative;
-  display: flex;
-  width: 400px;
-  height: 56.8px;
-  margin-top: 100px;
-  margin-bottom: 20px;
-  border: 1.6px solid #e95464;
-  border-radius: 15px;
-}
-
-.search {
-  width: 100px;
-  height: 30px;
-  color: white;
-  background-color: #e95464;
-  border: 0;
-  border-radius: 15px;
-}
-
-.search:hover {
-  background-color: #e84656;
-}
-
-.listItem {
-  padding: 20px 0;
-  border-bottom: 1px solid #ccc;
-}
-
-.listTitle {
-  margin: 0;
-  font-size: 1.5em;
-}
-
-.listDescription {
-  margin: 5px 0;
-  color: #555;
-}
